Add a fetch action for oVirt CA certs in provider Add/Edit
Add a new "fetch the engine CA certificate" link to the Add/Edit
provider modal.  Using an enhanced `/get-certificate` server route,
the CA cert is fetched and populated to the form field.

The `/get-certificate` server route is enhanced to be able to fetch
the CA certificate from an oVirt engine host.  Response data will
match the response of the `vsphere` request, and is used to populate
the certificate upload form field.

Manual upload of the CA certificate is still possible.